requirements.txt
python-telegram-bot==20.7
anthropic==0.34.2
python-dotenv==1.0.0
aiofiles==23.2.1
fastapi==0.104.1
uvicorn==0.24.0
jinja2==3.1.2
aiosqlite==0.19.0
